requirements.txt
neo4j==5.15.0
pandas==2.1.4
lightrag==0.1.0b6
openai==1.51.2
numpy==1.24.3
python-dotenv==1.0.0